+++ R/code-of-conduct.R
@@ -7,6 +7,12 @@
 #' <https://contributor-covenant.org>, version 1:
 #' <https://contributor-covenant.org/version/1/0/0>.
 #'
+#' If your package is going to CRAN the link to the CoC in your README must
+#' be an absolute link to a rendered website as `CODE_OF_CONDUCT.md` is not
+#' included in the package sent to CRAN. `use_code_of_conduct()` will
+#' automatically generate this link if you use pkgdown and have set the
+#' `url` field in `pkgdown.yml`; otherwise you'll need to fill it in yourself.
+#'
 #' @param path Path of the directory to put `CODE_OF_CONDUCT.md` in, relative to
 #'   the active project. Passed along to [use_directory()]. Default is to locate
 #'   at top-level, but `.github/` is also common.
@@ -24,10 +30,17 @@ use_code_of_conduct <- function(path = NULL) {
     ignore = is_package() && is.null(path)
   )
 
+  href <- project_pkgdown_url()
+  if (is.null(href)) {
+    ui_todo("Couldn't find pkgdown website; make sure to adjust block below")
+    href <- "{INSERT WEBSITE HERE}"
+  }
+  href <- paste0(href, "/CODE_OF_CONDUCT.html")
+
   ui_todo("Don't forget to describe the code of conduct in your README:")
   ui_code_block("
     Please note that the {ui_value(project_name())} project is released with a
-    [Contributor Code of Conduct]({save_as}).
+    [Contributor Code of Conduct]({href}).
     By contributing to this project, you agree to abide by its terms."
   )